misc refactor and typing
